# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Free Fire Esports Tournament API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments/{id}/registrations:
    get:
      operationId: listRegistrationsForTournament
      tags: [registrations]
      summary: Get registrations for a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
    post:
      operationId: registerForTournament
      tags: [registrations]
      summary: Register for a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{id}/approve:
    patch:
      operationId: approveRegistration
      tags: [registrations]
      summary: Approve a registration (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/{id}/reject:
    patch:
      operationId: rejectRegistration
      tags: [registrations]
      summary: Reject a registration (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
...
components:
  schemas:
...
    Tournament:
      type: object
      required: [id, name, mode, status, startDate, maxSlots, prizePool, entryFee]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        mode:
          type: string
          enum: [solo, duo, squad]
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        maxSlots:
          type: integer
        filledSlots:
          type: integer
        prizePool:
          type: number
        entryFee:
          type: number
        bannerUrl:
          type: ["string", "null"]
        roomId:
          type: ["string", "null"]
        roomPassword:
          type: ["string", "null"]
        countdownTo:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Registration:
      type: object
      required: [id, tournamentId, status, createdAt]
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        userId:
          type: string
        teamId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        freefireUid:
          type: string
        playerName:
          type: string
        paymentScreenshot:
          type: ["string", "null"]
        createdAt:
          type: string
        tournament:
          $ref: "#/components/schemas/Tournament"
        team:
          $ref: "#/components/schemas/Team"
    RegistrationInput:
      type: object
      required: [freefireUid, playerName]
      properties:
        freefireUid:
          type: string
          minLength: 1
        playerName:
          type: string
          minLength: 1
        teamId:
          type: integer
        paymentScreenshot:
          type: string
    Team:
      type: object
      required: [id, name, captainId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        tag:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        captainId:
          type: string
        totalWins:
          type: integer
        totalKills:
          type: integer
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
...
    TeamMember:
      type: object
      required: [id, userId, teamId, role, status]
      properties:
        id:
          type: integer
        userId:
          type: string
        teamId:
          type: integer
        role:
          type: string
          enum: [captain, member]
        status:
          type: string
          enum: [pending, active]
        freefireUid:
          type: ["string", "null"]
        playerName:
          type: ["string", "null"]
...