# .migration-backup/lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Free Fire Esports Tournament API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationStatus } from './registrationStatus';
import type { Team } from './team';
import type { Tournament } from './tournament';

export interface Registration {
  id: number;
  tournamentId: number;
  userId?: string;
  /** @nullable */
  teamId?: number | null;
  status: RegistrationStatus;
  freefireUid?: string;
  playerName?: string;
  /** @nullable */
  paymentScreenshot?: string | null;
  createdAt: string;
  tournament?: Tournament;
  team?: Team;
}