# .migration-backup/lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Free Fire Esports Tournament API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 1 */
  freefireUid: string;
  /** @minLength 1 */
  playerName: string;
  teamId?: number;
  paymentScreenshot?: string;
}